pyproject.toml
[project]
name = "lab4-cnn"
version = "0.1.0"
description = "Lab 4 CNN Colorization - DSAA2012"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "accelerate>=0.20.0",
    "datasets>=2.0.0",
    "diffusers>=0.15.0",
    "matplotlib>=3.5.0",
    "pillow>=9.0.0",
    "scikit-image>=0.19.0",
    "tqdm>=4.65.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
